# Optimize VectorX.Clamp according to proposed change