Closes #34017