@EgorBo 2020-04-13
Are `Vector4.Max\Min` actually vectorizes?
Your impl's codegen:
```asm
; Method Clamp1
G_M60623_IG01:
       sub      rsp, 56
       vzeroupper 
       vmovaps  qword ptr [rsp+20H], xmm6
       vmovaps  qword ptr [rsp+10H], xmm7
G_M60623_IG02:
       vmovss   xmm0, dword ptr [rdx]
       vmovss   xmm1, dword ptr [rdx+4]
       vmovss   xmm2, dword ptr [rdx+8]
       vmovss   xmm3, dword ptr [rdx+12]
       vmovss   xmm4, dword ptr [r8]
       vmovss   xmm5, dword ptr [r8+4]
       vmovss   xmm6, dword ptr [r8+8]
       vmovss   xmm7, dword ptr [r8+12]
       vucomiss xmm0, xmm4
       ja       SHORT G_M60623_IG04
G_M60623_IG03:
       jmp      SHORT G_M60623_IG05
G_M60623_IG04:
       vmovaps  xmm4, xmm0
G_M60623_IG05:
       vucomiss xmm1, xmm5
       ja       SHORT G_M60623_IG07
G_M60623_IG06:
       jmp      SHORT G_M60623_IG08
G_M60623_IG07:
       vmovaps  xmm5, xmm1
G_M60623_IG08:
       vucomiss xmm2, xmm6
       ja       SHORT G_M60623_IG10
G_M60623_IG09:
       jmp      SHORT G_M60623_IG11
G_M60623_IG10:
       vmovaps  xmm6, xmm2
G_M60623_IG11:
       vucomiss xmm3, xmm7
       ja       SHORT G_M60623_IG13
G_M60623_IG12:
       jmp      SHORT G_M60623_IG14
G_M60623_IG13:
       vmovaps  xmm7, xmm3
G_M60623_IG14:
       lea      rax, bword ptr [rsp]
       vmovss   dword ptr [rax], xmm4
       vmovss   dword ptr [rax+4], xmm5
       vmovss   dword ptr [rax+8], xmm6
       vmovss   dword ptr [rax+12], xmm7
       lea      rax, bword ptr [rsp]
       vmovss   xmm0, dword ptr [rax]
       vmovss   xmm1, dword ptr [rax+4]
       vmovss   xmm2, dword ptr [rax+8]
       vmovss   xmm3, dword ptr [rax+12]
       vmovss   xmm4, dword ptr [r9]
       vmovss   xmm5, dword ptr [r9+4]
       vmovss   xmm6, dword ptr [r9+8]
       vmovss   xmm7, dword ptr [r9+12]
       vucomiss xmm4, xmm0
       ja       SHORT G_M60623_IG16
G_M60623_IG15:
       jmp      SHORT G_M60623_IG17
G_M60623_IG16:
       vmovaps  xmm4, xmm0
G_M60623_IG17:
       vucomiss xmm5, xmm1
       ja       SHORT G_M60623_IG19
G_M60623_IG18:
       jmp      SHORT G_M60623_IG20
G_M60623_IG19:
       vmovaps  xmm5, xmm1
G_M60623_IG20:
       vucomiss xmm6, xmm2
       ja       SHORT G_M60623_IG22
G_M60623_IG21:
       jmp      SHORT G_M60623_IG23
G_M60623_IG22:
       vmovaps  xmm6, xmm2
G_M60623_IG23:
       vucomiss xmm7, xmm3
       ja       SHORT G_M60623_IG25
G_M60623_IG24:
       jmp      SHORT G_M60623_IG26
G_M60623_IG25:
       vmovaps  xmm7, xmm3
G_M60623_IG26:
       vmovss   dword ptr [rcx], xmm4
       vmovss   dword ptr [rcx+4], xmm5
       vmovss   dword ptr [rcx+8], xmm6
       vmovss   dword ptr [rcx+12], xmm7
       mov      rax, rcx
G_M60623_IG27:
       vmovaps  xmm6, qword ptr [rsp+20H]
       vmovaps  xmm7, qword ptr [rsp+10H]
       add      rsp, 56
       ret      
; Total bytes of code: 265
```
At the same time:
```csharp
public static Vector128<float> Clamp2(Vector128<float> value1,
    Vector128<float> min, Vector128<float> max)
{
    return Sse.Min(Sse.Max(value1, min), max);
}
```
Emits:
```asm
; Method Clamp2
G_M5676_IG01:
       vzeroupper 
G_M5676_IG02:
       vmovupd  xmm0, xmmword ptr [rdx]
       vmaxps   xmm0, xmm0, xmmword ptr [r8]
       vminps   xmm0, xmm0, xmmword ptr [r9]
       vmovupd  xmmword ptr [rcx], xmm0
       mov      rax, rcx
G_M5676_IG03:
       ret      
; Total bytes of code: 25
```