@AlFasGD 2020-04-13
@tannergooding would need to take a look at this, I'm not familiar with any method to further improve the codegen, I would assume that the proposed version would be just fine for this purpose, generating the best possible code.