@EgorBo 2020-04-13
Looks like none of the `[Intrinsics]` APIs in `Vector4_intrinsics.cs` actually accelerated or recognized by jit's importer.