@EgorBo 2020-04-13
Looks like `Vector4` vectorization is broken.