@JimBobSquarePants 2020-04-13
I'm seeing this on [sharplab](https://sharplab.io/#v2:C4LghgzgtgPgAgJgIwFgBQcAMACOSB0AcgK5QCmATgJYDGEA3OunAMy4LYDC2A3utgNxs4AFmwBZABQBKXv0EBfeQOVDcSAGzYAamRrAA9hTGcANmCgAHJJN36jYgB4AaHXsPHsUKgDtXdjzEoMEdpVUEI7ABeAD43e2N8cV9bdwckkMkXL19pV2DQxjQlNHQgA=) with the default settings.

``` asm
C.Clamp1(System.Numerics.Vector4, System.Numerics.Vector4, System.Numerics.Vector4)
    L0000: vzeroupper
    L0003: vmovupd xmm0, [esp+0x24]
    L0009: vmovupd xmm1, [esp+0x14]
    L000f: vmaxps xmm0, xmm0, xmm1
    L0013: vmovupd xmm1, [esp+0x4]
    L0019: vminps xmm0, xmm0, xmm1
    L001d: vmovupd [ecx], xmm0
    L0021: ret 0x30
```

We also saw a 5x improvement in ImageSharp when benchmarking the changes that led to the issue being raised.

https://github.com/SixLabors/ImageSharp/pull/1152/files#diff-c0645b9893e87cbacc35363b366826b1R55-R60