@EgorBo 2020-04-13
> I'm seeing this on [sharplab](https://sharplab.io/#v2:C4LghgzgtgPgAgJgIwFgBQcAMACOSB0AcgK5QCmATgJYDGEA3OunAMy4LYDC2A3utgNxs4AFmwBZABQBKXv0EBfeQOVDcSAGzYAamRrAA9hTGcANmCgAHJJN36jYgB4AaHXsPHsUKgDtXdjzEoMEdpVUEI7ABeAD43e2N8cV9bdwckkMkXL19pV2DQxjQlNHQgA=) with the default settings.

sharplab uses some stable .net core 3.0 runtime and I suspect something is broken in master (but maybe I am wrong)