@EgorBo 2020-04-13
Ah, I was using an old version of `System.Numerics.Vectors.dll` where `Vector4` type is not marked with `[Intrinsic]` so the master-runtime didn't recognize it as a SIMD.
Now the codegen is expected:
```asm
G_M52350_IG01:
       C5F877               vzeroupper
                                                ;; bbWeight=1    PerfScore 1.00
G_M52350_IG02:
       C5F91002             vmovupd  xmm0, xmmword ptr [rdx]
       C5F851C0             vsqrtps  xmm0, xmm0
       C5F91101             vmovupd  xmmword ptr [rcx], xmm0
       488BC1               mov      rax, rcx
                                                ;; bbWeight=1    PerfScore 18.25
G_M52350_IG03:
       C3                   ret
                                                ;; bbWeight=1    PerfScore 1.00

; Total bytes of code 19
```