@EgorBo 2020-04-13
Also, probably makes sense to optimize `Lerp` (e.g. using fma)