@tannergooding 2020-04-13
@EgorBo, there are a lot of optimizations to still be had. As I mentioned in #34017, I'm happy to take codegen improvements for any of the System.Numerics methods (provided the appropriate tests still pass and there are benchmarks).
The DirectX Math library has a lot of methods we don't expose and a lot of optimizations (for both ARM and x86): https://github.com/microsoft/DirectXMath/blob/master/Inc/DirectXMathVector.inl#L5586