@tannergooding 2020-04-13
Ideally we would also add some perf tests to https://github.com/dotnet/performance/blob/master/src/benchmarks/micro/libraries/System.Numerics.Vectors/Perf_Vector4.cs (and Vector2/3 if updating as well) and provide some minimal before/after numbers to help track the perf improvement.

Would you be willing to do that as part of this?

If so, https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-dotnet-runtime.md contains the workflow for testing a benchmark against local dotnet/runtime changes. @adamsitnik is also a good resource for any questions that might come up.