@tannergooding 2020-04-22
@AlFasGD, I think this is just missing some basic perf numbers from before/after the change so we can show it is a win.
It's needed largely just for tracking purposes to show we did our due diligence before merging.