@tannergooding 2020-04-23
Numbers look good:
### Before: 
```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.778 (1909/November2018Update/19H2)
AMD Ryzen 9 3900X, 1 CPU, 24 logical and 12 physical cores
.NET Core SDK=5.0.100-preview.4.20222.23
  [Host]     : .NET Core 5.0.0 (CoreCLR 5.0.20.22107, CoreFX 5.0.20.22107), X64 RyuJIT
  Job-BACAGA : .NET Core 5.0.0 (CoreCLR 5.0.20.22107, CoreFX 5.0.20.22107), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Runtime=.NET Core 5.0  Arguments=/p:DebugType=portable
Toolchain=netcoreapp5.0  IterationTime=250.0000 ms  MaxIterationCount=20
MinIterationCount=15  WarmupCount=1

|                 Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------:|------:|------:|----------:|
| Vector2.ClampBenchmark | 1.641 ns | 0.0316 ns | 0.0296 ns | 1.622 ns | 1.613 ns | 1.697 ns |     - |     - |     - |         - |
| Vector3.ClampBenchmark | 2.958 ns | 0.0235 ns | 0.0208 ns | 2.952 ns | 2.923 ns | 3.009 ns |     - |     - |     - |         - |
| Vector4.ClampBenchmark | 4.056 ns | 0.0742 ns | 0.0579 ns | 4.047 ns | 3.969 ns | 4.161 ns |     - |     - |     - |         - |
```

### After
```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.778 (1909/November2018Update/19H2)
AMD Ryzen 9 3900X, 1 CPU, 24 logical and 12 physical cores
.NET Core SDK=5.0.100-preview.2.20176.6
  [Host]     : .NET Core 5.0.0 (CoreCLR 5.0.20.22107, CoreFX 5.0.20.22107), X64 RyuJIT
  Job-JIKNIA : .NET Core 5.0 (CoreCLR 42.42.42.42424, CoreFX 42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:DebugType=portable  Toolchain=CoreRun
IterationTime=250.0000 ms  MaxIterationCount=20  MinIterationCount=15
WarmupCount=1

|                 Method |      Mean |     Error |    StdDev |    Median |       Min |       Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |----------:|----------:|----------:|----------:|----------:|----------:|------:|------:|------:|----------:|
| Vector2.ClampBenchmark | 0.2977 ns | 0.0020 ns | 0.0018 ns | 0.2975 ns | 0.2953 ns | 0.3015 ns |     - |     - |     - |         - |
| Vector3.ClampBenchmark | 0.7103 ns | 0.0123 ns | 0.0115 ns | 0.7069 ns | 0.6934 ns | 0.7312 ns |     - |     - |     - |         - |
| Vector4.ClampBenchmark | 0.2991 ns | 0.0046 ns | 0.0043 ns | 0.2972 ns | 0.2949 ns | 0.3080 ns |     - |     - |     - |         - |
```

I just followed the workflow listed under: https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-dotnet-runtime.md 
The process is basically to run `.\scripts\benchmarks_ci.py --frameworks netcoreapp5.0 --filter System.Numerics.Tests.*Clamp*` and then to run again passing `--coreroot` with the path to your local build of the PR

Vector3 isn;t "as good" because it needs to additional work to load the 3 elements, where-as Vector2 and Vector4 can each be a single read/write. The codegen looks fine in other scenarios where the JIT is able to better optimize.